AI Experiences Controller
closes LLMA-26
flag=none

This adds a controller for the AI Experience Model
in Canvas. The implemented operations are:
  - Create
  - Update
  - Delete
  - Index

We are not yet doing:
  - New
  - Edit
  - Show

For now, all we need is the ablilty to create/update
and delete AI Experiences. We will use the Index page
to display all active ones. The user will use the
index page to navigate to the create/edit page for
AI Experiences

Test Plan:
- Tests pass in Jenkins
- Manually hit the endpoints using curl or Postman
  - the url structure is this
  - {base_url}/courses/{:id}/ai_experiences/action/{:id}

Change-Id: I1b0116067de2c8c23b5a3ed9abd2ea2266b60ad6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/390459
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Caleb Guanzon <cguanzon@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Jason Anderson <jason.anderson@instructure.com>