support RSA InstAccess tokens
This supports the AMS being able to issue InstAccess tokens. The
existing symmetric key implementation contains a quirk making it
incompatible with other JWT libraries; this compatibility is required
since the AMS-API is not a Ruby project. The RSA portion of the
library does not have this quirk, and we just need to be able to pass
in the JWK information to configure it. Storing only a public key here
also has other advantages.

closes AG-1879
flag = none

test plan:
- Make no changes to the Canvas vault, and verify that it comes up OK.
- Verify that the vault contents have still been used to set up the
  InstAccess gem config via the rails console:
  `InstAccess.config.service_jwks`
- Add RSA public key material to the vault, restart, and verify that
  it comes up OK.
- Verify that the new key has been added to the InstAccess gem via the
  same rails console check.

Change-Id: I161c4083ba80b64b11b8242aa4e6c453fbbd1af7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/393049
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Griffin Zody <griffin.zody@instructure.com>
QA-Review: Griffin Zody <griffin.zody@instructure.com>
Product-Review: Griffin Zody <griffin.zody@instructure.com>