Create a way for JS code to send flash to next page
Closes FOO-5896
flag=none

When a form submission causes Rails to render a new page,
there’s a mechanism for Rails to pass along flash alerts to
the new page.  However, as we move more to React and InstUI
for submit actions, often the submit is an API call and then
the front end code has to do a page replacement on its own.

So, we need a way to communicate “… and also throw up this
flash alert when the new page loads” across page replacement
operations. sessionStorage should be a good candidate for
this.

This implements that functionality via `addFlashNoticeForNextPage`
and then uses it on UncrosslistForm.tsx to restore the original
Rails form-submit behavior of throwing up a "success" flash
on the new page after decrosslisting, indicating that the
decrosslisting was successful.

Also spruced up TS typing of props sent by data- attributes
in app/views/sections/show.html.erb to match the types of
the props in Rails.

Test plan:
* Have two courses with sections and crosslist a section from
  one course to the other
* Then de-crosslist the section from the other course
* You should get a new page load back at the original section
  as before, but now there should also be a "Success" flash
  alert on the original page saying that the de-crosslisting
  was successful

Change-Id: I4fc8da254b31ccfb045a2af54a20c8cc58530655
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/392458
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
QA-Review: Maya Tyner <maya.tyner@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>