Show Asset Reports on teacher-viewed grades page
This is the page that shows grades for a specific student. We already
show reports on this page when the viewer is a student, but the rules
about which reports should be shown to teachers are a little different.

To maintain consistency between old and new grades pages (and also
speedgrader really), I also needed to check the submissionType of the
submission, otherwise we will show "No result" for unsubmitted
submissions. This wasn't noticeable previously for students since we would
not show anything in the Document Processors column if there were no
reports at all (and there couldn't be any reports if there was no
submission, except for partially submitted discussions; this will be
fixed in INTEROP-9995)

closes INTEROP-9580
flag=lti_asset_processor

Test plan:
- have a course with at least one regular assignment with asset processor(s)
- in one of the regular assignments, have a student with an attempt, and make a report for that submission's asset with priority > 0.
- as a teacher, visit the grades page for the student (e.g.
  http://canvas-web.inseng.test/courses/1/grades/3)
- make sure the regular assignment shows "Needs Attention" in the document processor
- make a new attempt for the student, and make sure that it now shows "No result" for that assignment
- create a report for that attempt with priority 0. make sure it shows "All good"
- create a report for that attempt with priority 1 and some processingProgress other than "Processed". make sure it shows up as "Needs attention"
- make a graded discussion with an asset processor. make a discussion entry with the student but no reports yet.
- then make sure that discussion entry shows up on the grades page with "No result" in the asset processor
- make a report with priority 0 for the discussion entry and make sure it shows up as "All good"
- as a student edit the discussion entry. go back to the grade page as a teacher make sure the discussion entry now shows up as "No result"
- make a report with priority > 0 for the discussion entry and processingProgress other than "Processed". make sure it now shows up as "Needs Attention".
- make sure clicking the All Good or Needs Attention as appropriate opens the modal with the report details
- repeat these steps with the other version (new vs old) of the grades page.

Change-Id: I6608e66e3fc2f95d16464714f9b1821b284fa4a5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394101
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Csaba Csuzdi <csaba.csuzdi@instructure.com>
QA-Review: Csaba Csuzdi <csaba.csuzdi@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>