Ensure that only symbols are given as named arguments
The issue is that when using the ** operator on an object to pass arguments in to a method, if any of the keys in the object are
strings, it will throw an error about an unknown argument.

In some places, we do something like
`default_params.deep_merge(more_params)`, and `deep_merge`
stringifies the keys. So we need to make sure that they are
switched back to symbols before splatting them into method arguments.

Change-Id: I7e30c54a37ee7941bf32cc45df324745baaec58d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394054
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>