Fix quiz plannable_date null
https://gerrit.instructure.com/c/canvas-lms/+/392455
left out a change to the quiz type, which led to the
plannable_date for certain quizzes to be nil, which then
led to the planner not loading all of the future dates.
This simply updates the quiz plannable type to match the
rest of the changes.

closes LX-3530
flag=none

test plan:
- need a student in a section in a course
- create multiple plannable items in the near future and
  near past (assignments, discussions, etc)
- create a quiz in the near future and assign it only to the
  section, include a due date
- act as the student and go to the planner dashboard - ensure that
  you can see items in the future, including the quiz

Change-Id: I39463e6d559dea8f8ee4f6d116103836337f2ab9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/394852
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>
Product-Review: Sarah Gerard <sarah.gerard@instructure.com>